requirements.txt
duckdb>=1.2.0
chromadb>=0.6.3
crewai>=0.102.0
langchain>=0.3.19
sentence-transformers>=3.4.1
uuid>=1.30
python-dateutil>=2.9.0.post0
pandas>=2.2.3
tqdm>=4.67.1
langchain-openai>=0.3.7
langchain-community>=0.3.18
python-dotenv>=1.0.1